+++ app/services/title_extractor/hackernews_title_extractor.rb
@@ -1,0 +1,7 @@
+module TitleExtractor
+  class HackernewsTitleExtractor < Base
+    def title
+      "Hacker News Best"
+    end
+  end
+end